hw_6 fix

+++ hw_6/src/Tasks/delete.php
@@ -4,12 +4,16 @@ require __DIR__ . '/../../vendor/autoload.php';
 
 use Zayts\Hw6\Models\Task;
 
-$userId = $_GET['user_id'];
+$isAuth = \Zayts\Hw6\Auth::check();
+if (!$isAuth) {
+    return redirect("/login.php");
+}
+
 $taskId = $_GET['task_id'];
 
 $task= Task::find($taskId);
 $task->delete();
 
-header("Location: http://127.0.0.1:9000/index.php?user_id=$userId");
+redirect("/index.php");
